+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Abstract SyncTransforms function
+
 ## [3.0.0] - 2023-04-03
 ### Removed
 - Check for Angle Limit